lab-01.2-api-vs-selfhosted/requirements-api.txt
Flask==2.3.3
openai==0.28.1
PyYAML==6.0.1
pydantic==1.10.12
requests==2.31.0
